immigrate tokenconvert test from economicSystem.test to token convert
