+++ test/AElf.Contracts.TokenConverter.Tests/ACS1_ImplementTest.cs
@@ -18,7 +18,7 @@ namespace AElf.Contracts.TokenConverter
         [Fact]
         public async Task ChangeMethodFeeController_Test()
         {
-            await InitializeParliamentContractAsync();
+            //await InitializeParliamentContractAsync();
             var createOrganizationResult =
                 await ParliamentContractStub.CreateOrganization.SendAsync(
                     new CreateOrganizationInput
@@ -56,7 +56,7 @@ namespace AElf.Contracts.TokenConverter
         [Fact]
         public async Task ChangeMethodFeeController_WithoutAuth_Test()
         {
-            await InitializeParliamentContractAsync();
+            //await InitializeParliamentContractAsync();
             var createOrganizationResult =
                 await ParliamentContractStub.CreateOrganization.SendAsync(
                     new CreateOrganizationInput
